Add DM inbox item swipe gestures, add full size DM editor option
Resolves #1327
Resolves #1219
