+++ src/features/inbox/SendMessageBox.tsx
@@ -3,13 +3,17 @@ import { MaxWidthContainer } from "../shared/AppContent";
 import TextareaAutosize, {
   TextareaAutosizeProps,
 } from "react-textarea-autosize";
-import { IonIcon } from "@ionic/react";
-import { KeyboardEvent, useCallback, useState } from "react";
+import { IonButton, IonIcon } from "@ionic/react";
+import { KeyboardEvent, useCallback, useContext, useState } from "react";
 import useClient from "../../helpers/useClient";
 import useAppToast from "../../helpers/useAppToast";
 import { receivedMessages } from "./inboxSlice";
 import { useAppDispatch } from "../../store";
-import { arrowUp } from "ionicons/icons";
+import { expand, send as sendIcon } from "ionicons/icons";
+import { privateMessageSendFailed } from "../../helpers/toastMessages";
+import { css } from "@linaria/core";
+import { PageContext } from "../auth/PageContext";
+import { Person } from "lemmy-js-client";
 
 const MaxSizeContainer = styled(MaxWidthContainer)`
   height: 100%;
@@ -19,36 +23,36 @@ const SendContainer = styled.div`
   position: relative;
 
   padding: 0.5rem;
-
-  background: var(
-    --ion-tab-bar-background,
-    var(--ion-background-color-step-50, #f7f7f7)
-  );
-  border-top: 1px solid
-    var(
-      --ion-tab-bar-border-color,
-      var(
-        --ion-border-color,
-        var(--ion-background-color-step-150, rgba(0, 0, 0, 0.2))
-      )
-    );
 `;
 
 const InputContainer = styled.div`
   position: relative;
 
   display: flex;
+  align-items: flex-end;
+  gap: 4px;
 `;
 
 const Input = styled(TextareaAutosize)`
-  border: 1px solid var(--ion-color-medium);
+  border: 0.55px solid
+    var(
+      --ion-tab-bar-border-color,
+      var(
+        --ion-border-color,
+        var(
+          --ion-color-step-150,
+          var(--ion-background-color-step-150, rgba(0, 0, 0, 0.2))
+        )
+      )
+    );
   border-radius: 1rem;
 
   // Exact px measurements to prevent
   // pixel rounding browser inconsistencies
-  padding: 8px 30px 8px 1rem;
+  padding: 8px 12px;
   line-height: 18px;
   font-size: 16px;
+  margin: 0;
 
   background: var(--ion-background-color);
   color: var(--ion-text-color);
@@ -59,33 +63,32 @@ const Input = styled(TextareaAutosize)`
   appearance: none;
 `;
 
-const SendButton = styled(IonIcon)`
-  position: absolute;
-  bottom: calc(36px / 2);
-  transform: translateY(50%);
-  right: 4px;
-  background: var(--ion-color-primary);
-  height: 21px;
-  width: 21px;
-  border-radius: 50%;
-  padding: 3px;
-  color: white;
+const IconButton = styled(IonButton)`
+  margin: 0;
+  min-height: 36px;
+
+  ion-icon {
+    font-size: 22px;
+  }
 `;
 
 interface SendMessageBoxProps {
-  recipientId: number;
+  recipient: Person;
   onHeightChange?: TextareaAutosizeProps["onHeightChange"];
+  scrollToBottom?: () => void;
 }
 
 export default function SendMessageBox({
-  recipientId,
+  recipient,
   onHeightChange,
+  scrollToBottom,
 }: SendMessageBoxProps) {
   const dispatch = useAppDispatch();
   const [loading, setLoading] = useState(false);
   const [value, setValue] = useState("");
   const client = useClient();
   const presentToast = useAppToast();
+  const { presentCommentReply } = useContext(PageContext);
 
   const send = useCallback(async () => {
     setLoading(true);
@@ -95,13 +98,10 @@ export default function SendMessageBox({
     try {
       message = await client.createPrivateMessage({
         content: value,
-        recipient_id: recipientId,
+        recipient_id: recipient.id,
       });
     } catch (error) {
-      presentToast({
-        message: `Message failed to send. Please try again`,
-        color: "danger",
-      });
+      presentToast(privateMessageSendFailed);
       setLoading(false);
       throw error;
     }
@@ -110,7 +110,9 @@ export default function SendMessageBox({
 
     setLoading(false);
     setValue("");
-  }, [client, dispatch, presentToast, recipientId, value]);
+
+    scrollToBottom?.();
+  }, [client, dispatch, presentToast, recipient, value, scrollToBottom]);
 
   const onKeyDown = useCallback(
     (e: KeyboardEvent<HTMLTextAreaElement>) => {
@@ -126,6 +128,23 @@ export default function SendMessageBox({
     <SendContainer>
       <MaxSizeContainer>
         <InputContainer>
+          <IconButton
+            shape="round"
+            fill="clear"
+            onClick={() =>
+              presentCommentReply({ private_message: { recipient }, value })
+            }
+            aria-label="Open fullsize editor"
+          >
+            <IonIcon
+              icon={expand}
+              slot="icon-only"
+              onClick={send}
+              className={css`
+                transform: scale(1.1);
+              `}
+            />
+          </IconButton>
           <Input
             disabled={loading}
             placeholder="Message"
@@ -136,9 +155,21 @@ export default function SendMessageBox({
             onKeyDown={onKeyDown}
             onHeightChange={onHeightChange}
           />
-          {value.trim() && !loading && (
-            <SendButton icon={arrowUp} onClick={send} />
-          )}
+          <IconButton
+            disabled={!value.trim() || loading}
+            shape="round"
+            fill="clear"
+            onClick={send}
+            aria-label="Send message"
+          >
+            <IonIcon
+              icon={sendIcon}
+              slot="icon-only"
+              className={css`
+                transform: rotate(270deg);
+              `}
+            />
+          </IconButton>
         </InputContainer>
       </MaxSizeContainer>
     </SendContainer>
