--- src/features/shared/sliding/BaseSlidingVote.tsx
@@ -1,390 +1,0 @@
-import { styled } from "@linaria/react";
-import {
-  arrowDownSharp,
-  arrowUndo,
-  arrowUpSharp,
-  bookmark,
-  chevronCollapse,
-  chevronDown,
-  chevronExpand,
-  chevronUp,
-  eyeOffOutline,
-  eyeOutline,
-  mailOpen,
-  mailUnread,
-  share as shareIcon,
-} from "ionicons/icons";
-import React, { useCallback, useContext, useMemo } from "react";
-import SlidingItem, { ActionList, SlidingItemAction } from "./SlidingItem";
-import {
-  CommentReplyView,
-  CommentView,
-  PersonMentionView,
-  PostView,
-} from "lemmy-js-client";
-import store, { useAppDispatch, useAppSelector } from "../../../store";
-import { savePost, voteOnPost } from "../../post/postSlice";
-import {
-  postLocked,
-  replyStubError,
-  saveSuccess,
-} from "../../../helpers/toastMessages";
-import {
-  saveComment,
-  toggleCommentCollapseState,
-  voteOnComment,
-} from "../../comment/commentSlice";
-import { PageContext } from "../../auth/PageContext";
-import { SwipeAction, SwipeActions } from "../../../services/db";
-import useCollapseRootComment from "../../comment/inTree/useCollapseRootComment";
-import { getInboxItemId, markRead } from "../../inbox/inboxSlice";
-import { CommentsContext } from "../../comment/inTree/CommentsContext";
-import useAppToast from "../../../helpers/useAppToast";
-import { share } from "../../../helpers/lemmy";
-import { scrollCommentIntoViewIfNeeded } from "../../comment/inTree/CommentTree";
-import { AppContext } from "../../auth/AppContext";
-import { getCanModerate } from "../../moderation/useCanModerate";
-import { isStubComment } from "../../comment/CommentHeader";
-import { getVoteErrorMessage } from "../../../helpers/lemmyErrors";
-import { VOTE_COLORS } from "../../settings/appearance/themes/votesTheme/VotesTheme";
-
-const StyledItemContainer = styled.div`
-  --ion-item-border-color: transparent;
-`;
-
-type SlideableItem =
-  | CommentView
-  | PostView
-  | PersonMentionView
-  | CommentReplyView;
-
-function isInboxItem(
-  item: SlideableItem,
-): item is PersonMentionView | CommentReplyView {
-  if ("person_mention" in item) return true;
-  if ("comment_reply" in item) return true;
-  return false;
-}
-
-interface BaseSlidingVoteProps {
-  children: React.ReactNode;
-  className?: string;
-  item: SlideableItem;
-  rootIndex?: number;
-  collapsed?: boolean;
-  actions: SwipeActions;
-  onHide?: () => void;
-}
-
-export default function BaseSlidingVote(props: BaseSlidingVoteProps) {
-  const disableSwipes = useAppSelector(
-    (state) =>
-      state.gesture.swipe.disableLeftSwipes &&
-      state.gesture.swipe.disableRightSwipes,
-  );
-
-  if (!disableSwipes) {
-    return <BaseSlidingVoteInternal {...props} />;
-  } else {
-    // don't initialize all the sliding stuff if it's completely unused
-    return (
-      <StyledItemContainer className={props.className}>
-        {props.children}
-      </StyledItemContainer>
-    );
-  }
-}
-
-function BaseSlidingVoteInternal({
-  children,
-  className,
-  item,
-  rootIndex,
-  collapsed,
-  actions,
-  onHide,
-}: BaseSlidingVoteProps) {
-  const { presentLoginIfNeeded, presentCommentReply } = useContext(PageContext);
-  const { prependComments } = useContext(CommentsContext);
-
-  const { activePageRef } = useContext(AppContext);
-
-  const presentToast = useAppToast();
-  const dispatch = useAppDispatch();
-
-  const votesTheme = useAppSelector(
-    (state) => state.settings.appearance.votesTheme,
-  );
-
-  const postVotesById = useAppSelector((state) => state.post.postVotesById);
-  const commentVotesById = useAppSelector(
-    (state) => state.comment.commentVotesById,
-  );
-  const typedMyVote = item.my_vote as 1 | -1 | 0 | undefined;
-  const isPost = "unread_comments" in item;
-  const currentVote = isPost
-    ? (postVotesById[item.post.id] ?? typedMyVote)
-    : (commentVotesById[item.comment.id] ?? typedMyVote);
-
-  const postSavedById = useAppSelector((state) => state.post.postSavedById);
-  const commentSavedById = useAppSelector(
-    (state) => state.comment.commentSavedById,
-  );
-
-  const isHidden = useAppSelector(
-    (state) => state.post.postHiddenById[item.post?.id]?.hidden,
-  );
-
-  const readByInboxItemId = useAppSelector(
-    (state) => state.inbox.readByInboxItemId,
-  );
-
-  const disableLeftSwipes = useAppSelector(
-    (state) => state.gesture.swipe.disableLeftSwipes,
-  );
-  const disableRightSwipes = useAppSelector(
-    (state) => state.gesture.swipe.disableRightSwipes,
-  );
-
-  const onVote = useCallback(
-    async (score: 1 | -1 | 0) => {
-      if (presentLoginIfNeeded()) return;
-
-      if (isInboxItem(item)) dispatch(markRead(item, true));
-
-      try {
-        if (isPost) await dispatch(voteOnPost(item.post.id, score));
-        else await dispatch(voteOnComment(item.comment.id, score));
-      } catch (error) {
-        presentToast({
-          color: "danger",
-          message: getVoteErrorMessage(error),
-        });
-
-        throw error;
-      }
-    },
-    [presentLoginIfNeeded, isPost, dispatch, item, presentToast],
-  );
-
-  const reply = useCallback(async () => {
-    if (presentLoginIfNeeded()) return;
-
-    if (isInboxItem(item)) dispatch(markRead(item, true));
-
-    const canModerate = getCanModerate(item.community);
-
-    // Prevent replying to a comment that's been deleted, or removed by mod (if you're not a mod)
-    if (!isPost) {
-      const comment =
-        store.getState().comment.commentById[item.comment.id] ?? item.comment;
-
-      const stub = isStubComment(comment, canModerate);
-
-      if (stub) {
-        presentToast(replyStubError);
-        return;
-      }
-    }
-
-    if (item.post.locked && !canModerate) {
-      presentToast(postLocked);
-      return;
-    }
-
-    const reply = await presentCommentReply(item);
-    if (!isPost && reply) prependComments([reply]);
-  }, [
-    item,
-    isPost,
-    presentCommentReply,
-    presentLoginIfNeeded,
-    prependComments,
-    presentToast,
-    dispatch,
-  ]);
-
-  const { id, isSaved } = useMemo(() => {
-    if (isPost) {
-      const id = item.post.id;
-      return { id: id, isSaved: postSavedById[id] };
-    } else {
-      const id = item.comment.id;
-      return { id: id, isSaved: commentSavedById[id] };
-    }
-  }, [item, isPost, postSavedById, commentSavedById]);
-
-  const save = useCallback(async () => {
-    if (presentLoginIfNeeded()) return;
-    try {
-      await dispatch((isPost ? savePost : saveComment)(id, !isSaved));
-
-      if (!isSaved) presentToast(saveSuccess);
-    } catch (error) {
-      presentToast({
-        message: "Failed to mark item as saved",
-        color: "danger",
-      });
-      throw error;
-    }
-  }, [presentLoginIfNeeded, dispatch, isPost, id, isSaved, presentToast]);
-
-  const saveAction = useMemo(() => {
-    return {
-      icon: bookmark,
-      trigger: save,
-      bgColor: "success",
-      slash: isSaved,
-    };
-  }, [save, isSaved]);
-
-  const hideAction = useMemo(() => {
-    return onHide
-      ? {
-          icon: isHidden ? eyeOutline : eyeOffOutline,
-          trigger: () => {
-            if (presentLoginIfNeeded()) return;
-            onHide();
-          },
-          bgColor: isHidden ? "tertiary" : "danger",
-        }
-      : undefined;
-  }, [presentLoginIfNeeded, isHidden, onHide]);
-
-  const collapseRootComment = useCollapseRootComment(
-    !isPost ? item : undefined,
-    rootIndex,
-  );
-  const collapseToTopAction = useMemo(() => {
-    return collapseRootComment
-      ? {
-          icon: collapsed ? chevronExpand : chevronCollapse,
-          trigger: collapseRootComment,
-          bgColor: "tertiary",
-        }
-      : undefined;
-  }, [collapsed, collapseRootComment]);
-
-  const collapse = useCallback(
-    (e: TouchEvent | MouseEvent) => {
-      if (isPost) return;
-
-      dispatch(toggleCommentCollapseState(item.comment.id));
-
-      if (e.target) scrollCommentIntoViewIfNeeded(e.target, activePageRef);
-    },
-    [dispatch, isPost, item, activePageRef],
-  );
-  const collapseAction = useMemo(() => {
-    return {
-      icon: collapsed ? chevronDown : chevronUp,
-      trigger: collapse,
-      bgColor: "tertiary",
-    };
-  }, [collapsed, collapse]);
-
-  const isRead = useMemo(() => {
-    return isInboxItem(item) ? readByInboxItemId[getInboxItemId(item)] : false;
-  }, [item, readByInboxItemId]);
-
-  const markUnread = useCallback(async () => {
-    if (!isInboxItem(item)) return;
-
-    try {
-      await dispatch(markRead(item, !isRead));
-    } catch (error) {
-      presentToast({
-        message: "Failed to mark item as unread",
-        color: "danger",
-      });
-      throw error;
-    }
-  }, [dispatch, presentToast, item, isRead]);
-
-  const markUnreadAction = useMemo(() => {
-    return {
-      icon: isRead ? mailUnread : mailOpen,
-      trigger: markUnread,
-      bgColor: "primary-fixed",
-    };
-  }, [markUnread, isRead]);
-
-  const shareTrigger = useCallback(async () => {
-    share(isPost ? item.post : item.comment);
-  }, [isPost, item]);
-
-  const allActions: Record<SwipeAction, SlidingItemAction | undefined> =
-    useMemo(() => {
-      return {
-        none: undefined,
-        upvote: {
-          icon: arrowUpSharp,
-          trigger: () => {
-            onVote(currentVote === 1 ? 0 : 1);
-          },
-          bgColor: VOTE_COLORS.UPVOTE[votesTheme],
-          slash: currentVote === 1,
-        },
-        downvote: {
-          icon: arrowDownSharp,
-          trigger: () => {
-            onVote(currentVote === -1 ? 0 : -1);
-          },
-          bgColor: VOTE_COLORS.DOWNVOTE[votesTheme],
-          slash: currentVote === -1,
-        },
-        reply: {
-          icon: arrowUndo,
-          trigger: reply,
-          bgColor: "primary-fixed",
-        },
-        save: saveAction,
-        hide: hideAction,
-        "collapse-to-top": collapseToTopAction,
-        collapse: collapseAction,
-        "mark-unread": markUnreadAction,
-        share: {
-          icon: shareIcon,
-          trigger: shareTrigger,
-          bgColor: "primary-fixed",
-        },
-      };
-    }, [
-      currentVote,
-      reply,
-      saveAction,
-      hideAction,
-      collapseAction,
-      collapseToTopAction,
-      markUnreadAction,
-      onVote,
-      shareTrigger,
-      votesTheme,
-    ]);
-
-  const startActions: ActionList = useMemo(
-    () =>
-      !disableLeftSwipes
-        ? [allActions[actions.start], allActions[actions.farStart]]
-        : [undefined, undefined],
-    [disableLeftSwipes, allActions, actions],
-  );
-
-  const endActions: ActionList = useMemo(
-    () =>
-      !disableRightSwipes
-        ? [allActions[actions.end], allActions[actions.farEnd]]
-        : [undefined, undefined],
-    [disableRightSwipes, allActions, actions],
-  );
-
-  return (
-    <SlidingItem
-      startActions={startActions}
-      endActions={endActions}
-      className={className}
-    >
-      {children}
-    </SlidingItem>
-  );
-}
